bashfulbytes.com/requirements.txt
Beaker==1.8.0
blinker==1.4
chardet==2.3.0
cloudpickle==0.2.1
colorama==0.3.7
decorator==4.0.9
docutils==0.12
ghp-import2==1.0.1
gunicorn==19.4.5
husl==4.0.3
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
Logbook==0.12.5
lxml==3.6.0
Markdown==2.6.6
MarkupSafe==0.23
natsort==5.0.1
oauthlib==1.0.3
packaging==16.7
Pillow==3.2.0
psutil==4.1.0
Pygments==2.1.3
pyinotify==0.9.6
pyparsing==2.1.4
PyRSS2Gen==1.1
python-dateutil==2.5.3
pyxdg==0.25
PyYAML==3.11
virtualenv==15.0.1
webassets==0.11.1
Werkzeug==0.11.3
Yapsy==1.11.223
zope.interface==4.1.3
